:sparkles: Use `gdate` when necessary
On BSD-related environment (mostly related to Mac OS) the default
`date` does not point to the GNU date version that this script
expects. So, we'll check if the available `date` command is GNU date,
if that's not the case we'll try to use `gdate`, from `coreutils`, and
raise an exception if none of them is available. This behaviour was
reported in #1

Signed-off-by: Antonio Ossa Guerra <aaossa@uc.cl>